Don't build with flat chrome.
The original need for that was to allow to override content from other
packages, but in the end this happened to be made possible with
extensions and proper overlays.

Also, build-depend on zip for jar chrome. This has the side effect to
make nsIPref.java removal from MozillaInterfaces-src.jar work.