Use application.ini from where the xulrunner-stub lies
Whether it is a symlink or not, use the application.ini file from where
the xulrunner-stub is, if present. If not, fallback to upstream method
of following symlinks. This will avoid copying the stub in all xulrunner
applications like upstream does.

Squashed:
Refactored the previous change allowing xulrunner-stub to be symlinked

It will now work in more cases, including not calling the stub from
$PATH or an absolute location.