Implement a test application and run it during the test rule
Test both with and without libjemalloc, forcing memory allocation libraries
with LD_PRELOAD to avoid relying on what is linked at build time, as it may
change in the future.
If we built against libjemalloc and the test fails, automatically rebuild
without.
Also add build dependencies on xvfb, xfonts-base and xauth to be able to
launch xulrunner tests.