Properly launch applications set in $HOME/.mailcap
The problem is that the GNOME handler is launched despite what is set in
$HOME/.mailcap. This is due to nsMIMEInfoUnix being a mix between GNOME
and mailcap entries, and that GNOME entries have some magic behaviour.

We just change this by having mailcap entries be handled by "dumb"
nsMIMEInfoImpl, while GNOME entries are handled by nsMIMEInfoUnix.

At the same time, we do some cleanup in handlers for protocols, that just
happened to work because of some side effects of the current code, which
could have become prone to unexpected breakage with future changes.