Don't rely on readelf -p in dh_xulrunner
Now rely on objdump -h, which should be available on stable and probably
oldstable. This means we can lift the binutils dependencies.

Note there was a binutils build dependency before, for a very old mips
bug, but the required version was already in oldstable, so removing the
build-dep altogether just works.